プレイヤーの操作で、加速度の調整、衝突時のバグ修正
