+++ Assets/Prefabs/Katsumata.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 442c18beaaffaae41b727f971e1699c0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
