+++ Assets/Prefabs/Katsumata/PlayerBoard.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ab42b3ef5acd0694998a29df74156a0c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
