+++ Assets/Scenes/Katumata/Materials/Floor.mat
@@ -61,9 +61,9 @@ Material:
     - _DetailNormalMapScale: 1
     - _DstBlend: 0
     - _GlossMapScale: 1
-    - _Glossiness: 0.85
+    - _Glossiness: 0.493
     - _GlossyReflections: 1
-    - _Metallic: 0.88
+    - _Metallic: 0
     - _Mode: 0
     - _OcclusionStrength: 1
     - _Parallax: 0.02
@@ -73,5 +73,5 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 1, g: 1, b: 1, a: 1}
-    - _EmissionColor: {r: 0.3301887, g: 0.31617126, b: 0.31617126, a: 1}
+    - _Color: {r: 0.8773585, g: 0.8773585, b: 0.8773585, a: 1}
+    - _EmissionColor: {r: 0.3207547, g: 0.26441082, b: 0.246618, a: 1}
