+++ Assets/Scenes/Katumata/Physics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5ef0d68f403e31248aeacd7a3d473a6a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
