+++ Assets/Scenes/Katumata/Physics/Friction.physicMaterial.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ca52facecf3ba7c4c9d5f01827a1f4ab
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
